Set json library version to the one Starsector bundles (#1)
Using the wrong one can lead to odd errors at runtime due to missing
methods. In particular, it seems that the bundled version doesn't have
`JSONObject.keySet`. The exact version is still somewhat of a mystery,
but it's definitely something earlier than 2013 - 2013+ have `keySet`.
The files in the bundled `json.jar` are timestamped with 9th of May,
2010, so 2009 seems like a reasonable guess.